+++ icat/data.py
@@ -18,10 +18,10 @@ import panel as pn
 import param
 
 import icat
-from icat import _kill_param_auto_docstring
 from icat.anchors import TFIDFAnchor
 from icat.instance import InstanceViewer
 from icat.table import TableContentsTemplate
+from icat.utils import _kill_param_auto_docstring
 
 _kill_param_auto_docstring()