+++ icat/view.py
@@ -11,9 +11,9 @@ import pandas as pd
 import panel as pn
 from ipyanchorviz import AnchorViz
 
-from icat import _kill_param_auto_docstring
 from icat.anchors import Anchor, DictionaryAnchor, SimilarityFunctionAnchor, TFIDFAnchor
 from icat.histograms import Histograms
+from icat.utils import _kill_param_auto_docstring
 
 _kill_param_auto_docstring()