+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Added
+* Useful imports to top level module.
+
+
+
+
 ## [0.2.0] - 2023-07-05
 
 ### Added
@@ -24,12 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-
 ## [0.1.2] - 2023-04-17
 
 ### Fixed
 * More missing dependencies
-